MAVEDB-68 ORCID login no longer works.
Our test site has not changed recently, but logging in no longer works there. This suggested a change in ORCID’s OpenID Connect (OIDC) service.

The POST request to ORCID that initiates sign-in includes response_type and scope parameters. These are configured when we create the OIDC Vue client. Until now, they have been set as follows:

{
  response_type: 'id_token'
  scope: 'openid profile'
}

They appear in the POST request’s query string as response_type=id_token&scope=openid+profile.

It seems that ORCID’s OpenID service now wants these parameters to be set differently, as response_type=token&scope=openid.

The scope parameter can be changed in the OIDC Vue client configuration.

But if we change the response_type parameter to token in the client configuration, the underlying oidc-client-ts library will no longer expect an id_token property in the response, and logging in will fail. So we need to keep this parameter as is, but change the query string in the sign-in URL.

To achieve this, I’ve monkey-patched the oidc-client-ts library. I consider this a temporary solution. A longer-term solution will involve finding or writing a new OIDC client.

As an alternative, I evaluated the ORCID widget (https://github.com/ORCID/orcid-auth-widget), which has not been affected by whatever ORCID changed caused this. While it seems desirable to use ORCID’s own login client, the code isn’t really suitable for use in a modern single-page application.