Allow query strings in collection URLs defined by item types.
The item type mechanism used here is derived from another project where client-side item types corresponded directly to REST resources and collections. The items Vuex store module depends on this way of defining item types.

We will need to use the items Vuex store module for fetching suggested identifiers that will be shown in a text input with auto-completion. In this case, the item types for the three external identifier databases all refer to the same REST API endpoint, which needs to be called with query parameters. Therefore, we need to allow the REST API endpoint configuration for an item type to include query parameters, which will be merged with any additional parameters applied to the items module.