External target gene identifiers
This update provides the user interface for attaching external gene identifiers and offsets to a score set's target gene. Please see the corresponding mavedb-api change set for more information about this change.

Notice that we do not perform client-side identifier validation or normalization for Ensembl, RefSeq, or UniProt identifiers; we rely on server-side validation. But client-side validation can be added in src/lib/identifiers.js if desired.