Add External Publication Identifier Search Item
Adds the `external-publication-identifier-search` item type which sends
a POST request with some search text to the MaveDB API. The endpoint
responds with `ExternalPublication` objects for every publication
database that contains a publication matching the provided search text.

Note that if a publication identifier does not pass validation for a
database server, we do not attempt to retrieve a publication from it.