Allow Users to Associate Pre-prints with Experiments and Scoresets
Allows users to associate pre-print articles from the bioRxiv and medRxiv servers to
new experiments and scoresets. Improves the UI elements associated with selecting these
publications so that users are able to see more publication metadata than just the
publication identifier.

For identifiers within our database, we provided Autocomplete results based on a matching
substring. When a user types an identifier that is valid in an external database but that
we do not have in our database, we fetch critical metadata to provide in our autocomplete
from this external database. This API result is mediated by the MaveDB API server.