Create ScoreSetTable Layout
The default scoreset table view present in the search screen will be useful in
a future publication identifiers screen. This change preps for that addition
by splitting the default scoreset table shown in the search screen into a new
importable component.

This table should be wrapped inside a flexbox with a div for it to be displayed
properly.