Fixed the taxonomy dropdown problem.
The problem seems to be that after the dropdown button is clicked, the AutoComplete control used as a dropdown stays in "searching" mode until the list of suggestions changes. If the search text has not changed, we do not run a new API query to populate the list of suggestions, so the AutoComplete remains in this mode.