Allow Score Set Filtration via Target Accession
Adds a new filter group to the search screen that allows score sets to
be filtered by target accession numbers (if defined).

This change also fixes a bug that occurred on attempting to access the
length attribute of undefined values. These values are now set to an
empty array by default.