Add Highlights View Component
Adds a Highlights View Vue component usable by other Vue components within the application.

This component takes prop: `model`, which may be either `Target`, `ScoreSet`, or `Experiment`. Depending
on which model is passed to this component, it will display a leaderboard of properties defined on the
passed model and/or a set of charts to show the proportional representation of certain model properties.

In the case of ScoreSets and Experiments, these properties are Keywords, Publication Identifiers, and
DOI Identifiers. In the case of Targets, these properties are accessions, gene names, UniProt, RefSeq,
and Ensembl Identifiers. The Target highlight view will also show the proportional representation of
the target organism, reference, accession, and gene category fields.

Leaderboards are populated as the user moves between them and data is retained in the browser until
the page is reloaded.

If additional properties are needed, they can be added as additional lines of configuration under the
appropriate model. If another model must be added, it should receive its own display v-if section in
this components' template as well as its own configuration options.