Use ProcessingState to Display Variant Processing Information on Unpublished Score Sets
Uses the `processingState` score set property and associated `processingErrors` property to display
information about the state of the variant processing job for Score Sets. Errors in the job are now
displayed in a more friendly table, replacing the toast notification that may have contained information
about hundred of validation failures.