Remove Any Lingering References to a 'Default' User Role
Removes lingering references to a default user role, replacing them with 'ordinary user'. This default
role was resulting in access keys trying to be created for a default user, which is not a real role, thus
raising 422s from the backend. Creating default API keys now hits the correct endpoint and creates the keys
successfully.