Allow Ter in the heatmap.
To date, the heatmap has only accepted the single-character code "*" for stop codons. This usage was present in some old score sets, but it is not valid MAVE-HGVS notation.

We will continue to support single-character notation for both amino acids and stop codons in the heatmap, but here we allow Ter as well. On the server side, single-character notation is not accepted, and the old score sets with invalid strings will be updated.