https://github.com/VariantEffect/mavedb-ui/issues/216: SVG/PNG Chart Exports
Adds a chart export lib module in support of this task. SVGs are handled with custom logic
to allow for grouped export (in the heatmap, the chart is composed of two SVGs grouped within
a scroll div). For PNGs, we use an external module as this conversion of composed SVGs was
giving us trouble during PNG Rasterization.