Use an accessor field to make selections on Heatmap and Histogram
Previously, the selection logic assumed that the provided datum would be the same object as the object present in the data array of the heatmap/histogram. With the separate loading of clinical controls and subsequent association with variants out of the heatmap/histogram context, it was necessary to instead make datum selections based on some pre-set accessor field on the provided datum. This makes it significantly easier to select datum objects in cases where we do not have a guarnatee of object equality between internal heatmap/histogram datums and their externally selected counterparts.