Updates for unspecified score range classifications
Adds support for unspecified score range classifications. As part of this change, a new string
formatter `beautifyString` was also added that converts underscores to spaces and makes the string
title case. This better supports score ranges, which are stored in the database without spaces and
in lower case, by making them appear in a more readable format.