Fixes #369: Update ScoreSetCreator and Editor to support object based external identifier setting
Previously, trying to add a external gene identifier that was not already present in our database would result in the front end silently ignoring the user input. This change removes the restriction that user input be selected from the autocomplete list and fixes the input so that it is in the format expected by the score set creation schema.