Prevent vertical scrolling on heatmap container
In some cases, the presence of heatmap tooltip causes the dynamic height to be calculated incorrectly, resulting in extra empty space below the heatmap and triggering the vertical scroll behavior. This change prevents vertical scrolling in all cases, relying on the scroll container height calculation to match the heatmap height.