Handle ScoreSetHeatmap row selection in ScoreSetVisualizer component
This change handles the selection of a scoreset row when the y-axis tick label is clicked to color the protein visualization by the scores for a particular amino acid. This selectionData prop passed to the visualization component now includes scores and colors for each row, in addition to the mean, min and max. A watcher has been added to the `colorBy` ref to switch between the different color schemes, which maps the selectionData prop values to the format expected by the PDBe-molstar viewer instance.