Add Footer Component
Adds a footer component with site logo and dynamic links. This footer is now
part of the site's default layout and appears at the bottom of each page. On
certain pages like score-set, experiment, and experiment-pages, the footer may
contain an extra link for reporting an issue with a dataset. To add other
links based on router context, follow the same format. These links will always
appear at the front of the link list. In addition, an Account link will be
included if a user is signed in.