Update heatmap to allow Y-axis to display row group labels if defined and enabled
Updates to allow rows to include group codes and labels, which can be used to display groups names in Y-axis alongside existing tick mark labels. This feature is disabled by default, but can be enabled via drawYGroups Accessor function. Requires `groupCode` and `groupLabel` to be included in rows data to be displayed.