Allow for heatmap row range selection and update protein visualization to display selected AA
To support the selection by y-axis tick label groups, adding functionality to select a range of rows to heatmap. This returns a 2D array containing all of the data from the selected rows, which can then be used to calculate averages or other aggregate statistics for display.

Also updates the protein visualization and parent component to include the selected amino acid name with the "color by" options on the SelectButton shown at the above the protein model, when a single row of the scoreset heatmap is selected. This allows user to toggle between coloring the protein model by the scores for that specific row and the other 3 options (mean, max. missense, and min. missense)