Use investigator-supplied functional class ranges to determine histogram colors.
This change correctly orients the red-white-blue color gradient according to normal and abnormal ranges specified by the investigator. It also handles multiple ranges of either type (and overlapping functional class ranges, which should not normally arise).

One perhaps less clearly intended consequence is that score sets with only one functional class range (normal or abnormal) get colored with a gradient from blue to white or red to white, with no use of the other functional class's color. When an abnormal range and baseline score are present, the baseline score could be used as a blue control point. But in other cases there is no obvious resolution.