Add a heatmap configuration prop governing coloring when there is only one functional class range (normal or abnormal).
Without this option, the new, more correct functional-class-sensitive coloring results in a gradient from blue to white or red to white when there is only a normal or abnormal class; the other functional class's color does not get used. In the past, though the orientation may have been incorrect with respect to the functional classes, both functional class colors would have been used. The new forceBothClassificationColors option causes the missing color to be introduced, as long as there is at least one data point below or above all existing range(s) of the other functional class, and not merely on the boundary.