Update heatmap and related Vue components to select row group selection and corresponding calculation for coloring protein structure
Updates to heatmap to replace recently added row range selection function with row group selection, in order to return the selected rows data as well as the group name and label values. This function is then utilized by ScoreSetHeatmap to pass the relevant data to ScoreSetVisualizer for calculating the mean values for the scores at each position across all rows of the group. The resuling mean scores are then used to assign colors in the ProteinStructureView component, and the group code is used as the label in the "Color by" select button to show the currently selected color used for the protein structure and allow the user to toggle between this and other color options.