Additional changes to ScoreSetCreator to parse files for, and include score and counts column metadata when saving
To match changes made in ScoreSetEditor for the same fields, ScoreSetCreator will now parse uploaded files for score and count column metadata and submit as part of datasetColumns field of new record.

Also cleans up some remaining references to `this.item`, which is only applicable for ScoreSetEditor and isn't used when creating new score sets.