fix: update parseScoresOrCounts function to handle namespaced parameter across components
While the `/variants/data` endpoint is namespaced, the `/scores` and `/counts` endpoints are not (since they represent the namespace directly. To make the CSV parser more generally, this change adds a parameter that controls whether to treat the CSV as namespaced. In cases where scores/counts column names contain a period, this prevents them from being improperly treated as namespaces.