Replace Alphafold URL HEAD request with check of PDBe Molstar viewer strucureRefMap
To catch and display UI errors related to invalid or otherwise unavailable AlphaFold CIF file URLs that are being used in PDBe Molstar viewer of ProteinStructureView component, the previous solution of making a HEAD request was not behaving consistently. In some cases, CORS errors were occuring, but no obvious pattern was observed.

To address this, the HEAD request is being replaced by a check of the size of the PDBe molstar viewer instance's structureRefMap property. This value should be greater than 0 if the protein structure loaded successfully, with an error message displayed otherwise.