Updating components and CSS post-upgrade to PrimeVue4 and tailwindCSS
Multiple pages and components were impacted by the upgrade to PriveVue 4. These changes are to help restore the look and feel of those prior to the upgrade.

This includes the use of the theme.css file that is specific to PrimeVue 3 to maintain consistency with CSS classes already in use throughout the application. Going forward, PrimeVue 3 specific class names should be phased out in favor of PrimeVue 4 Aura theme and tailwind CSS plus any additional custom classes that are needed.